Include last position in POSITION command
This allows us to send a `POSITION` command to inform instances that the
stream position has advanced without instances thinking they've missed
updates.

This is so that sharded event persisters can update their local
positions in the events stream to the lastest position and periodically
tell other instances about it.

This helps ensure that the calculated minimum persisted position of the
events stream advances even if only one event persister is writing.
Background processes that read off the event stream use the minimum
persisted position, and so keeping that up to date ensures that those
background processes don't get stuck.